Tests refactored
